requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
dacite==1.8.1
pymongo==4.8.0
python-dotenv==0.18.0
Flask-JWT-Extended==4.6.0
requests == 2.32.3